The IdentityProviderUserAuthoritiesProvider class was checking if the returned role Attribute's value was of String type or String Array type. However, the value was coming as a List and hence, the application was returning the user role as "USER" irrespective to the roles returned by the okta provider. Added a else if condition to handle the scenario where the role attribute value is of List type.
Renamed the CASRoleMappingConfig.java file to RoleAttributeNameMappingConfig.java

bean was not getting generated for IdentityProviderUserAuthoritiesProvider in the project-security.xml file of the application. (The check was whether it is a instance type of OpenIDRoleMappingConfig).

RADX-308 - Access denied to the pages/services when a user is trying to access a page/service even though that user has given the access
RADX-307 - Security info call response is not fetching the user roles properly

Change-Id: I02e9b194e39fe2b96c80795b9a3e19a416ec8a18